[FIX] standalone_viewport: stop following selection
All the viewports would listen to the selection events and scroll
to follow the selection, even though it only made sens for the main
viewport, not the carousel viewports.

Task: 6481678
